Move MemoryAware Passes to correct place
